Add options folder and options/utils
